addressservice/api/views.py
from cities_light.models import City
from rest_framework import viewsets

from api.models import Address
from api.serializers import AddressSerializer, CitySerializer


class CityViewSet(viewsets.ModelViewSet):
    queryset = City.objects.all()
    serializer_class = CitySerializer


class AddressViewSet(viewsets.ModelViewSet):
    queryset = Address.objects.all()
    serializer_class = AddressSerializer
