pyproject.toml
[tool.poetry]
name = "qogita-challenge"
version = "0.1.0"
description = ""
authors = ["Karl <kturrell@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
django-cities-light = "^3.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
